Analysis of Stock Prices from 2012 to 2018
In this notebook I will examine 3 stocks. Activision, Apple and Microsoft. Each stock will be represented as a time series starting from 2012 and ending at the end of 2018.

I will analyse :

The 9 and 30 day moving average of each stock
The volatility rate of each stock
The actual change in stock price
Compare the 3 stocks together and try find trends